% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BS_option.R
\name{BS_option}
\alias{BS_option}
\title{Black-Scholes option price, delta and gamma}
\usage{
BS_option(s, K, r, tau, sigma, CP)
}
\arguments{
\item{s}{double containing price of the underlying asset}

\item{K}{double containing the strike price of the option}

\item{r}{double containing the risk free interest rate}

\item{tau}{double containing the time until maturity (in years)}

\item{sigma}{double containing the implied volatility}

\item{CP}{string to determine whether option is european call, 'C', or put, 'P'}
}
\value{
An object of class "ExtremeClass"
}
\description{
Generates option price, delta and gamma using standard Black-Scholes framework
}
\examples{
BS_option(100, 100, 0.02, 1, 0.2, 'C')
}
