% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GBM.R
\name{plot.LessExtremeClass}
\alias{plot.LessExtremeClass}
\title{Plot method for class LessExtremeClass}
\usage{
\method{plot}{LessExtremeClass}(x, ...)
}
\arguments{
\item{x}{An object of class LessExtremeClass}

\item{...}{Additional arguments to be passed to print function (not used in this function)}
}
\value{
a plot
}
\description{
Plot method for class LessExtremeClass
}
\examples{
GBM <- simulate_GBM(100, 0.02, 0.02, 0.2, 1000, 52)

plot(GBM)

}
