% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_option.R
\name{print.ExtremeClass}
\alias{print.ExtremeClass}
\title{Print ExtremeClass objects}
\usage{
\method{print}{ExtremeClass}(x, ...)
}
\arguments{
\item{x}{Object of class "ExtremeClass"}

\item{...}{Additional arguments to be passed to print function (not used in this function)}
}
\value{
None (prints output directly)
}
\description{
This function prints the Black-Scholes options price, delta, and gamma of an ExtremeClass object.
}
\examples{
x <- BS_option(100, 100, 0.02, 1, 0.2, "C")

print(x)
}
