% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_option.R
\name{summary.ExtremeClass}
\alias{summary.ExtremeClass}
\title{Summarize an ExtremeClass object}
\usage{
\method{summary}{ExtremeClass}(object, ...)
}
\arguments{
\item{object}{An object of class ExtremeClass}

\item{...}{Additional arguments to be passed to print function (not used in this function)}
}
\value{
A list containing summary information about the ExtremeClass object
}
\description{
This function computes and presents a summary of an ExtremeClass object,
including the parameters used to calculate the Black-Scholes option price, delta, and gamma.
}
\examples{
# Create an example ExtremeClass object
x <- BS_option(100, 100, 0.02, 1, 0.2, 'C')

# Summarize the ExtremeClass object
summary(x)

}
