% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_GBM.R
\name{simulate_GBM}
\alias{simulate_GBM}
\title{Simulate Geometric Brownian Motion}
\usage{
simulate_GBM(S0, r, mu, sigma, num_simulations, num_steps)
}
\arguments{
\item{S0}{an initial value of the GBM}

\item{r}{the risk free interest rate}

\item{mu}{the drift component}

\item{sigma}{the volatility component}

\item{num_simulations}{amount of simulations to create}

\item{num_steps}{amount of steps to simulate}
}
\value{
an object of class "LessExtremeClass"
}
\description{
Simulate Geometric Brownian Motion
}
\examples{
simulate_GBM(100, 0.02, 0.02, 0.2, 1000, 52)
}
